Added training data and changed datatypes

+++ onboarding/Michael/model.py
@@ -28,10 +28,10 @@ class FractionMLP:
 
         1 hidden layer for now
         """
-        self.W1 = torch.randn(input_size, hidden_size, requires_grad=True)
-        self.b1 = torch.randn(1, hidden_size, requires_grad=True)
-        self.W2 = torch.randn(hidden_size, output_size, requires_grad=True)
-        self.b2 = torch.randn(1, output_size, requires_grad=True)
+        self.W1 = torch.randn(input_size, hidden_size, requires_grad=True, dtype=torch.float32)
+        self.b1 = torch.randn(1, hidden_size, requires_grad=True, dtype=torch.float32)
+        self.W2 = torch.randn(hidden_size, output_size, requires_grad=True, dtype=torch.float32)
+        self.b2 = torch.randn(1, output_size, requires_grad=True, dtype=torch.float32)
 
     def forward(self, X):
         """
